requirements.txt
python-decouple==3.8
aiohttp==3.8.6